requirements.txt
flask
flask-cors
python-nmap
requests